#1212 refactoring: move proxy checks from `DownloadFile` to `DownloadFileWithProxyServer`
