+++ src/test/java/com/codeborne/selenide/commands/DownloadFileTest.java
@@ -1,12 +1,12 @@
 package com.codeborne.selenide.commands;
 
-import com.codeborne.selenide.Browser;
-import com.codeborne.selenide.Config;
 import com.codeborne.selenide.Driver;
 import com.codeborne.selenide.DriverStub;
 import com.codeborne.selenide.SelenideConfig;
+import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.files.FileFilter;
 import com.codeborne.selenide.files.FileFilters;
+import com.codeborne.selenide.impl.DownloadFileToFolder;
 import com.codeborne.selenide.impl.DownloadFileWithHttpRequest;
 import com.codeborne.selenide.impl.DownloadFileWithProxyServer;
 import com.codeborne.selenide.impl.WebElementSource;
@@ -14,7 +14,6 @@ import com.codeborne.selenide.proxy.SelenideProxyServer;
 import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
 import java.io.File;
@@ -31,11 +30,13 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 class DownloadFileTest implements WithAssertions {
-  private final Config config = mock(Config.class);
+  private final SelenideConfig config = new SelenideConfig();
   private final Driver driver = mock(Driver.class);
   private final DownloadFileWithHttpRequest httpget = mock(DownloadFileWithHttpRequest.class);
   private final DownloadFileWithProxyServer proxy = mock(DownloadFileWithProxyServer.class);
-  private final DownloadFile command = new DownloadFile(httpget, proxy);
+  private final DownloadFileToFolder folder = mock(DownloadFileToFolder.class);
+  private final DownloadFile command = new DownloadFile(httpget, proxy, folder);
+  private final SelenideElement seLink = mock(SelenideElement.class);
   private final WebElementSource linkWithHref = mock(WebElementSource.class);
   private final WebElement link = mock(WebElement.class);
   private final File file = new File("some-file.yxy");
@@ -49,12 +50,11 @@ class DownloadFileTest implements WithAssertions {
 
   @Test
   void canDownloadFile_withHttpGetRequest() throws IOException {
-    when(config.proxyEnabled()).thenReturn(false);
-    when(config.fileDownload()).thenReturn(HTTPGET);
+    config.fileDownload(HTTPGET);
 
     when(httpget.download(any(), any(WebElement.class), anyLong(), any())).thenReturn(file);
 
-    File f = command.execute(null, linkWithHref, new Object[]{8000L});
+    File f = command.execute(seLink, linkWithHref, new Object[]{8000L});
 
     assertThat(f).isSameAs(file);
     verify(httpget).download(driver, link, 8000L, none());
@@ -63,48 +63,28 @@ class DownloadFileTest implements WithAssertions {
 
   @Test
   void canDownloadFile_withProxyServer() throws IOException {
-    SelenideConfig config = new SelenideConfig().proxyEnabled(true).fileDownload(PROXY);
+    config.proxyEnabled(true).fileDownload(PROXY);
     SelenideProxyServer selenideProxy = mock(SelenideProxyServer.class);
     when(linkWithHref.driver()).thenReturn(new DriverStub(config, null, null, selenideProxy));
-    when(proxy.download(any(), any(), any(), anyLong(), any())).thenReturn(file);
+    when(proxy.download(any(), any(), anyLong(), any())).thenReturn(file);
 
-    File f = command.execute(null, linkWithHref, new Object[]{9000L});
+    File f = command.execute(seLink, linkWithHref, new Object[]{9000L});
 
     assertThat(f).isSameAs(file);
-    verify(proxy).download(linkWithHref, link, selenideProxy, 9000L, none());
+    verify(proxy).download(linkWithHref, link, 9000L, none());
     verifyNoMoreInteractions(httpget);
   }
 
   @Test
-  void proxyServerShouldBeEnabled() {
-    when(config.proxyEnabled()).thenReturn(false);
-    when(config.fileDownload()).thenReturn(PROXY);
-
-    assertThatThrownBy(() -> command.execute(null, linkWithHref, new Object[0]))
-      .isInstanceOf(IllegalStateException.class)
-      .hasMessageContaining("Cannot download file: proxy server is not enabled");
-  }
-
-  @Test
-  void proxyServerShouldBeStarted() {
-    SelenideConfig config = new SelenideConfig().proxyEnabled(true).fileDownload(PROXY);
-    when(linkWithHref.driver()).thenReturn(new DriverStub(config, mock(Browser.class), mock(WebDriver.class), null));
-
-    assertThatThrownBy(() -> command.execute(null, linkWithHref, new Object[0]))
-      .isInstanceOf(IllegalStateException.class)
-      .hasMessageContaining("Cannot download file: proxy server is not started");
-  }
-
-  @Test
   void defaultTimeout() {
-    SelenideConfig config = new SelenideConfig().timeout(3L);
+    config.timeout(3L);
 
     assertThat(command.getTimeout(config, new Object[0])).isEqualTo(3L);
   }
 
   @Test
   void customTimeout() {
-    SelenideConfig config = new SelenideConfig().timeout(3L);
+    config.timeout(3L);
 
     assertThat(command.getTimeout(config, new Object[]{2L})).isEqualTo(2L);
     assertThat(command.getTimeout(config, new Object[]{2L, none()})).isEqualTo(2L);
